+++ lib/braintree/transaction_line_item.rb
@@ -1,18 +1,21 @@
 module Braintree
   class TransactionLineItem
     include BaseModule
-    attr_reader :quantity
-    attr_reader :name
+    attr_reader :commodity_code
     attr_reader :description
-    attr_reader :kind
-    attr_reader :unit_amount
-    attr_reader :unit_tax_amount
-    attr_reader :unit_of_measure
     attr_reader :discount_amount
+    attr_reader :image_url
+    attr_reader :kind
+    attr_reader :name
+    attr_reader :product_code
+    attr_reader :quantity
     attr_reader :tax_amount
     attr_reader :total_amount
-    attr_reader :product_code
-    attr_reader :commodity_code
+    attr_reader :unit_amount
+    attr_reader :unit_of_measure
+    attr_reader :unit_tax_amount
+    attr_reader :upc_code
+    attr_reader :upc_type
     attr_reader :url
 
     def initialize(gateway, attributes)